refactor: split tests and add test vectors
